requirements.txt
python-dotenv==1.0.1
openai==1.30.0
Pillow==10.1.0
mss==7.2.0
customtkinter
pyinstaller